Convert HK80 geographical coordinates to HK80 UTM coordinates

Convert HK80 geographical coordinates to HK80 UTM coordinates

HK80GEO_TO_HK80UTM(latitude, longitude)

Arguments

latitude
Numeric, latitude in decimal format.

longitude
Numeric, longitude in decimal format.

Details

Using equation (1) to (3) in the explanatory Note within this package. The document is available at : http://www.geodetic.gov.hk/smo/gsi/data/pdf/explanatorynotes.pdf

Value

A list including the HK80UTM Coordinates, including: N, E, and Zone.

References

Survey & Mapping Office Lands Department, Hong Kong Government (1995). Explanatory Notes on Geodetic Datums in Hong Kong, available at: http://www.geodetic.gov.hk/smo/gsi/data/pdf/explanatorynotes.pdf

Author

Jinlong Zhang

Note

The latitude and longitude input should be constrained to Hong Kong.

See also

Examples

options(digits = 15)
#### The example on the explanatory notes
HK80GEO_TO_HK80UTM(22 + 26/60 + 6.76/3600, 114 + 10/60 + 20.46/3600)

#>                  N                E zone
#> 1 2483772.48730165 208931.916380936   50

#### Output
##  $N
##  [1] 2483772.487
##  
##  $E
##  [1] 208931.9164
##  
##  $zone
##  [1] "50Q"

###### The results from the explanatory notes
#### 2483772, 208932

###### The similar conversion is not available 
###### from the online conversion tool.